cast DWMWA_WINDOW_CORNER_PREFERENCE to u32

+++ src-tauri/src/main.rs
@@ -248,7 +248,7 @@ fn start_capture(app: &AppHandle) {
                     let preference: u32 = DWMWCP_DONOTROUND as u32;
                     DwmSetWindowAttribute(
                         hwnd,
-                        DWMWA_WINDOW_CORNER_PREFERENCE,
+                        DWMWA_WINDOW_CORNER_PREFERENCE as u32,
                         &preference as *const u32 as *const std::ffi::c_void,
                         std::mem::size_of::<u32>() as u32,
                     );
